+++ packages/router/src/bindings/messaging/metaTxRequest.ts
@@ -9,8 +9,7 @@ import {
 import { BigNumber } from "ethers";
 
 import { NoTransactionId } from "../../lib/errors";
-import { TransactionReasons } from "../../lib/entities";
-import { incrementFees, incrementGasConsumed } from "../../lib/helpers";
+import { incrementFees } from "../../lib/helpers";
 import { getOperations } from "../../lib/operations";
 import { getContext } from "../../router";
 
@@ -70,7 +69,6 @@ export const metaTxRequestBinding = async (
       await messaging.publishMetaTxResponse(from, inbox, { chainId, transactionHash: tx.transactionHash });
       // Increment collected fees + gas used on relayer fee
       incrementFees(txData.receivingAssetId, txData.receivingChainId, BigNumber.from(relayerFee), requestContext);
-      incrementGasConsumed(txData.receivingChainId, tx.gasUsed, TransactionReasons.Relay, requestContext);
     }
     logger.info("Handled fulfill request", requestContext, methodContext);
   } finally {