+++ packages/router/src/lib/helpers/metrics.ts
@@ -1,11 +1,12 @@
 import { createLoggingContext, jsonifyError, RequestContext } from "@connext/nxtp-utils";
-import { constants, BigNumber, utils } from "ethers";
+import { constants, BigNumber, utils, providers } from "ethers";
 
 import { getContext } from "../../router";
 import {
   ExpressiveAssetBalance,
   feesCollected,
   gasConsumed,
+  relayerFeesPaid,
   totalTransferredVolume,
   TransactionReason,
 } from "../entities";
@@ -318,31 +319,90 @@ export const incrementFees = async (
   );
 };
 
+/**
+ * Increments gas consumed by the router signer each time it sends a transaction
+ *
+ * @notice This function should only be called through the `adapters/contract/contract.ts` functions. This is because
+ * costs paid for the transaction will be different if it is via a relayer or the txservice, and only those
+ * functions will have proper context into that.
+ * @param chainId - Chain transaction was sent on
+ * @param receipt - Receipt to calculate gas for
+ * @param reason - Why transaction was sent
+ * @param _requestContext - Request context for top-level method
+ * @returns void
+ */
 export const incrementGasConsumed = async (
   chainId: number,
-  gas: BigNumber,
+  receipt: providers.TransactionReceipt | undefined,
   reason: TransactionReason,
   _requestContext: RequestContext,
 ) => {
-  const { logger } = getContext();
+  if (!receipt) {
+    return;
+  }
+  const { logger, txService } = getContext();
 
   const { requestContext, methodContext } = createLoggingContext(incrementGasConsumed.name, _requestContext);
+  const { cumulativeGasUsed, effectiveGasPrice } = receipt;
+  const price = effectiveGasPrice ?? (await txService.getGasPrice(chainId, requestContext));
   logger.debug("Method start", requestContext, methodContext, {
     chainId,
-    gas: gas.toString(),
+    gas: cumulativeGasUsed.toString(),
+    price: price.toString(),
   });
 
-  const usd = await convertToUsd(constants.AddressZero, chainId, gas.toString(), requestContext);
+  const usd = await convertToUsd(
+    constants.AddressZero,
+    chainId,
+    cumulativeGasUsed.mul(price).toString(),
+    requestContext,
+  );
 
   logger.debug("Got gas fees in usd", requestContext, methodContext, {
     chainId,
-    gas: gas.toString(),
+    gas: cumulativeGasUsed.toString(),
+    price: price.toString(),
     usd: usd.toString(),
   });
 
   // Update counter
   // TODO: reason type
   gasConsumed.inc({ chainId, reason }, usd);
+};
+
+/**
+ * Increments relayer fees paid by the router signer each time it sends a transaction via relayers.
+ *
+ * @notice This function should only be called through the `adapters/contract/contract.ts` functions. This is because
+ * costs paid for the transaction will be different if it is via a relayer or the txservice, and only those
+ * functions will have proper context into that.
+ * @param chainId - Chain transaction was sent on
+ * @param relayerFee - Amount sent to relayer
+ * @param assetId - Asset used to pay relayer
+ * @param reason - Why transaction was sent
+ * @param _requestContext - Request context for top-level method
+ * @returns void
+ */
+export const incrementRelayerFeesPaid = async (
+  chainId: number,
+  relayerFee: string,
+  assetId: string,
+  reason: TransactionReason,
+  _requestContext: RequestContext,
+) => {
+  const { logger } = getContext();
+
+  const { requestContext, methodContext } = createLoggingContext(incrementTotalTransferredVolume.name, _requestContext);
+  logger.debug("Method start", requestContext, methodContext, {
+    chainId,
+    assetId,
+    relayerFee,
+    reason,
+  });
+
+  const usd = await convertToUsd(assetId, chainId, relayerFee, requestContext);
+
+  relayerFeesPaid.inc({ reason, chainId, assetId }, usd);
 };
 
 export const incrementTotalTransferredVolume = async (