+++ packages/router/test/adapters/contract/contract.spec.ts
@@ -17,6 +17,7 @@ import { Interface } from "ethers/lib/utils";
 import { BigNumber, constants } from "ethers";
 
 import * as SharedFns from "../../../src/lib/helpers/shared";
+import * as MetricsFns from "../../../src/lib/helpers/metrics";
 import * as ContractFns from "../../../src/adapters/contract/contract";
 import {
   prepareTransactionManager,
@@ -34,6 +35,7 @@ import { routerAddrMock, routerContractAddressMock } from "../../utils";
 import { routerAddress, messagingMock, signerAddress, txServiceMock } from "../../globalTestHook";
 import { SanitationCheckFailed } from "../../../src/lib/errors";
 import { ERC20Interface } from "@connext/nxtp-contracts/typechain/ERC20";
+import { TransactionReasons } from "../../../src/lib/entities";
 
 const requestContext = createRequestContext("TEST");
 const encodedDataMock = "0xabcde";
@@ -56,6 +58,8 @@ describe("Contract Adapter", () => {
   let isRouterWhitelistedStub: SinonStub;
   let isChainSupportedByGelatoStub: SinonStub;
   let gelatoSendStub: SinonStub;
+  let gasConsumedStub: SinonStub;
+  let relayerFeesPaidStub: SinonStub;
 
   beforeEach(() => {
     interfaceMock = createStubInstance(Interface);
@@ -76,6 +80,9 @@ describe("Contract Adapter", () => {
     isRouterWhitelistedStub = stub(SharedFns, "isRouterWhitelisted");
     isChainSupportedByGelatoStub = stub(ContractFns, "isChainSupportedByGelato").returns(true);
     gelatoSendStub = stub(ContractFns, "gelatoSend");
+
+    gasConsumedStub = stub(MetricsFns, "incrementGasConsumed");
+    relayerFeesPaidStub = stub(MetricsFns, "incrementRelayerFeesPaid");
   });
 
   afterEach(() => {
@@ -149,7 +156,7 @@ describe("Contract Adapter", () => {
     });
   });
 
-  describe("#prepare", () => {
+  describe("#prepareTransactionManager", () => {
     beforeEach(() => {
       sanitationStub = stub(SharedFns, "sanitationCheck").resolves();
     });
@@ -170,6 +177,8 @@ describe("Contract Adapter", () => {
         },
       ]);
       expect(res).to.deep.eq(txReceiptMock);
+      expect(relayerFeesPaidStub.callCount).to.be.eq(0);
+      expect(gasConsumedStub.calledOnceWithExactly(chainId, txReceiptMock, TransactionReasons.PrepareReceiver));
     });
 
     it("should fail if encoding fails", async () => {
@@ -180,7 +189,7 @@ describe("Contract Adapter", () => {
     });
   });
 
-  describe("#fulfill", () => {
+  describe("#fulfillTransactionManager", () => {
     beforeEach(() => {
       sanitationStub = stub(SharedFns, "sanitationCheck").resolves();
     });
@@ -199,6 +208,8 @@ describe("Contract Adapter", () => {
         },
       ]);
       expect(res).to.deep.eq(txReceiptMock);
+      expect(relayerFeesPaidStub.callCount).to.be.eq(0);
+      expect(gasConsumedStub.calledOnceWithExactly(chainId, txReceiptMock, TransactionReasons.FulfillSender));
     });
 
     it("should fail if sanitation check fails", async () => {
@@ -216,7 +227,7 @@ describe("Contract Adapter", () => {
     });
   });
 
-  describe("#cancel", () => {
+  describe("#cancelTransactionManager", () => {
     beforeEach(() => {
       sanitationStub = stub(SharedFns, "sanitationCheck").resolves();
     });
@@ -233,6 +244,8 @@ describe("Contract Adapter", () => {
         },
       ]);
       expect(res).to.deep.eq(txReceiptMock);
+      expect(relayerFeesPaidStub.callCount).to.be.eq(0);
+      expect(gasConsumedStub.calledOnceWithExactly(chainId, txReceiptMock, TransactionReasons.CancelSender));
     });
 
     it("should fail if sanitation check fails", async () => {
@@ -445,6 +458,8 @@ describe("Contract Adapter", () => {
         sigMock,
       ]);
       expect(res).to.deep.eq(txReceiptMock);
+      expect(relayerFeesPaidStub.callCount).to.be.eq(0);
+      expect(gasConsumedStub.calledOnceWithExactly(chainIdMock, txReceiptMock, TransactionReasons.PrepareReceiver));
       // Preflight estimate gas check should *not* be called if we aren't using relayers.
       expect(txServiceMock.getGasEstimate.callCount).to.be.eq(0);
       // Should have used txservice to send the tx.
@@ -474,6 +489,15 @@ describe("Contract Adapter", () => {
       );
 
       expect(res).to.deep.eq(txReceiptMock);
+      expect(gasConsumedStub.callCount).to.be.eq(0);
+      expect(
+        relayerFeesPaidStub.calledOnceWithExactly(
+          chainIdMock,
+          "1",
+          routerRelayerFeeAssetMock,
+          TransactionReasons.PrepareReceiver,
+        ),
+      );
 
       // Preflight estimate gas check should be called if we use relayers.
       expect(txServiceMock.getGasEstimate).to.have.been.calledOnceWithExactly(chainIdMock, onchainTxMock);
@@ -502,7 +526,15 @@ describe("Contract Adapter", () => {
       );
       expect(messagingMock.publishMetaTxRequest.callCount).to.be.eq(1);
       expect(res).to.deep.eq(txReceiptMock);
-
+      expect(gasConsumedStub.callCount).to.be.eq(0);
+      expect(
+        relayerFeesPaidStub.calledOnceWithExactly(
+          chainIdMock,
+          "1",
+          routerRelayerFeeAssetMock,
+          TransactionReasons.PrepareReceiver,
+        ),
+      );
       // Preflight estimate gas check should be called if we use relayers.
       expect(txServiceMock.getGasEstimate).to.have.been.calledOnceWithExactly(chainIdMock, onchainTxMock);
     });
@@ -584,6 +616,15 @@ describe("Contract Adapter", () => {
         requestContext,
       );
       expect(res).to.deep.eq(txReceiptMock);
+      expect(gasConsumedStub.callCount).to.be.eq(0);
+      expect(
+        relayerFeesPaidStub.calledOnceWithExactly(
+          chainIdMock,
+          "1",
+          routerRelayerFeeAssetMock,
+          TransactionReasons.FulfillSender,
+        ),
+      );
       // Preflight estimate gas check should be called if we use relayers.
       expect(txServiceMock.getGasEstimate).to.have.been.calledOnceWithExactly(chainIdMock, onchainTxMock);
     });
@@ -611,6 +652,15 @@ describe("Contract Adapter", () => {
       );
       expect(messagingMock.publishMetaTxRequest.callCount).to.be.eq(1);
       expect(res).to.deep.eq(txReceiptMock);
+      expect(gasConsumedStub.callCount).to.be.eq(0);
+      expect(
+        relayerFeesPaidStub.calledOnceWithExactly(
+          chainIdMock,
+          "1",
+          routerRelayerFeeAssetMock,
+          TransactionReasons.FulfillSender,
+        ),
+      );
       // Preflight estimate gas check should be called if we use relayers.
       expect(txServiceMock.getGasEstimate).to.have.been.calledOnceWithExactly(chainIdMock, onchainTxMock);
     });
@@ -661,6 +711,8 @@ describe("Contract Adapter", () => {
         sigMock,
       ]);
       expect(res).to.deep.eq(txReceiptMock);
+      expect(relayerFeesPaidStub.callCount).to.be.eq(0);
+      expect(gasConsumedStub.calledOnceWithExactly(chainIdMock, txReceiptMock, TransactionReasons.CancelSender));
       // Preflight estimate gas check should *not* be called if we aren't using relayers.
       expect(txServiceMock.getGasEstimate.callCount).to.be.eq(0);
       // Should have used txservice to send the tx.
@@ -689,6 +741,16 @@ describe("Contract Adapter", () => {
         requestContext,
       );
       expect(res).to.deep.eq(txReceiptMock);
+
+      expect(gasConsumedStub.callCount).to.be.eq(0);
+      expect(
+        relayerFeesPaidStub.calledOnceWithExactly(
+          chainIdMock,
+          "1",
+          routerRelayerFeeAssetMock,
+          TransactionReasons.CancelSender,
+        ),
+      );
       // Preflight estimate gas check should be called if we use relayers.
       expect(txServiceMock.getGasEstimate).to.have.been.calledOnceWithExactly(chainIdMock, onchainTxMock);
     });
@@ -716,7 +778,15 @@ describe("Contract Adapter", () => {
       );
       expect(messagingMock.publishMetaTxRequest.callCount).to.be.eq(1);
       expect(res).to.deep.eq(txReceiptMock);
-      // Preflight estimate gas check should be called if we use relayers.
+      expect(gasConsumedStub.callCount).to.be.eq(0);
+      expect(
+        relayerFeesPaidStub.calledOnceWithExactly(
+          chainIdMock,
+          "1",
+          routerRelayerFeeAssetMock,
+          TransactionReasons.CancelSender,
+        ),
+      );
       expect(txServiceMock.getGasEstimate).to.have.been.calledOnceWithExactly(chainIdMock, onchainTxMock);
     });