+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- [router] Update gas consumed, relayer fees tracking
+
 ## 0.1.12
 
 - [router] Fix metrics error